Minimization of the chat service: strip of all unnecessary functions
Unfortunately, due to time reasons, I won't be able to create a chat in the complexity and functionality that I have hoped for. For that reason the chat has been stripped of everything apart the most basic functionality. That does not mean it will stay like that forever. I hope that in the future I will find the time to implement everything that has been lost by this commit.